begin to deprec db.third_parties module
