+++ src/simmate/apps/configs.py
@@ -3,6 +3,10 @@
 from django.apps import AppConfig
 
 
+class AflowConfig(AppConfig):
+    name = "simmate.apps.aflow"
+
+
 class BadelfConfig(AppConfig):
     name = "simmate.apps.badelf"
 
@@ -15,12 +19,28 @@ class CleaseConfig(AppConfig):
     name = "simmate.apps.clease"
 
 
+class CodConfig(AppConfig):
+    name = "simmate.apps.cod"
+
+
 class EvolutionConfig(AppConfig):
     name = "simmate.apps.evolution"
 
 
+class JarvisConfig(AppConfig):
+    name = "simmate.apps.jarvis"
+
+
 class MaterialsProjectConfig(AppConfig):
     name = "simmate.apps.materials_project"
+
+
+class OqmdConfig(AppConfig):
+    name = "simmate.apps.oqmd"
+
+
+class QuantumEspressoConfig(AppConfig):
+    name = "simmate.apps.quantum_espresso"
 
 
 class RdkitConfig(AppConfig):
@@ -29,10 +49,6 @@ class RdkitConfig(AppConfig):
 
 class VaspConfig(AppConfig):
     name = "simmate.apps.vasp"
-
-
-class QuantumEspressoConfig(AppConfig):
-    name = "simmate.apps.quantum_espresso"
 
 
 class WarrenLabConfig(AppConfig):
