+++ src/simmate/apps/materials_project/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .structure import MatprojStructure
