+++ src/simmate/apps/oqmd/models.py/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
