--- src/simmate/database/third_parties/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .aflow import AflowPrototype, AflowStructure
-from .cod import CodStructure
-from .jarvis import JarvisStructure
-from .materials_project import MatprojStructure
-from .oqmd import OqmdStructure
-from .utilities import load_default_sqlite3_build, load_remote_archives
